+++ packages/react/src/components/DataTable/src/components/DataTableCell/dataTableCell.types.ts
@@ -1,0 +1,15 @@
+import { HTMLAttributes, ReactNode } from "react";
+
+import { TableCellProperties } from "@nimbus-ds/components";
+
+export interface DataTableCellProperties {
+  /**
+   * Content of the List component.
+   * @TJS-type React.ReactNode
+   */
+  children: ReactNode;
+}
+
+export type DataTableCellProps = DataTableCellProperties &
+  Omit<HTMLAttributes<HTMLElement>, "color"> &
+  TableCellProperties;