+++ packages/react/src/components/NavTabs/src/components/NavTabsItem/navTabsItem.spec.tsx
@@ -7,13 +7,16 @@ import { NavTabsItemProps } from "./navTabsItem.types";
 const iconElement = "Icon element";
 const mockedClickFunction = jest.fn();
 
-const makeSut = (rest: Omit<NavTabsItemProps, "icon" | "onClick">) => {
+const makeSut = (
+  rest: Omit<NavTabsItemProps, 'icon' | 'onClick' | 'ariaLabel'>
+) => {
   render(
     <NavTabsItem
       {...rest}
       icon={iconElement}
       data-testid="nav-tabs-item-element"
       onClick={mockedClickFunction}
+      ariaLabel="nav-tabs-item"
     />
   );
 };
@@ -23,6 +26,12 @@ describe("GIVEN <NavTabsItem />", () => {
     it("SHOULD render the icon element correctly", () => {
       makeSut({});
       expect(screen.getByText(iconElement)).toBeDefined();
+    });
+
+    it("SHOULD render the aria label correctly", () => {
+      makeSut({});
+      const button = screen.getByRole("button");
+      expect(button.getAttribute("aria-label")).toBe("nav-tabs-item");
     });
 
     it("SHOULD execute the onClick correctly", () => {